requirements.txt
Cython==0.29.16
h5py==2.9.0
numpy==1.18.2
scipy==1.4.1
Pillow==8.1.1
pycocotools==2.0
PyYAML==5.4
tensorboard==2.1.1
torch==1.4.0
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.44.0
yacs==0.1.6
